cli/requirements.txt
click>=6.7
tqdm
pdfminer
pdf2image
pandas 
scikit-learn
tabulate

# S3 access
boto3

# Grobid integration
requests

# used for testing
pytest

# Used for pdf-token extraction
pdfplumber

# Used for calculating scores 
pycocotools

# Used for OCR
pytesseract
google-cloud-vision
opencv-python
